Rename dimensions_helper to dimensions
